chore: extract byte layout to separate module
